add items api

+++ src/modules/items/items.service.ts
@@ -1,7 +1,8 @@
+import { Injectable, NotFoundException } from '@nestjs/common';
 import { InvoiceItem, User } from '@prisma/client';
 
 import { CreateItemDto } from './dto/create-item.dto';
-import { Injectable } from '@nestjs/common';
+import { NOT_FOUND } from '../../errors/errors.constants';
 import { PrismaService } from '../../providers/prisma/prisma.service';
 import { UpdateItemDto } from './dto/update-item.dto';
 
@@ -9,8 +10,18 @@ import { UpdateItemDto } from './dto/update-item.dto';
 export class ItemsService {
   constructor(private readonly prisma: PrismaService) {}
 
-  async create(createItemDto: CreateItemDto): Promise<InvoiceItem> {
-    const { name, description, price, enabled, invoiceId } = createItemDto;
+  async create(
+    createItemDto: CreateItemDto,
+    user: User,
+    invoiceId: number,
+  ): Promise<InvoiceItem> {
+    const { name, description, price, enabled } = createItemDto;
+
+    const invoice = await this.prisma.invoice.findUnique({
+      where: { id: invoiceId, userId: user.id },
+    });
+
+    if (!invoice) throw new NotFoundException(NOT_FOUND);
 
     const newItem = await this.prisma.invoiceItem.create({
       data: {
@@ -18,19 +29,41 @@ export class ItemsService {
         description,
         price,
         enabled,
-        Invoice: { connect: { id: invoiceId } },
+        Invoice: { connect: { id: invoice.id } },
       },
     });
 
     return newItem;
   }
 
-  async findAll(): Promise<InvoiceItem[]> {
-    const allItems = await this.prisma.invoiceItem.findMany({});
+  async findAll(invoiceId: number, userId: number): Promise<InvoiceItem[]> {
+    const invoice = await this.prisma.invoice.findUnique({
+      where: { id: invoiceId, userId },
+    });
+
+    if (!invoice) throw new NotFoundException(NOT_FOUND);
+
+    const allItems = await this.prisma.invoiceItem.findMany({
+      where: {
+        invoiceId: invoice.id,
+      },
+    });
     return allItems;
   }
 
-  async findOne(id: number): Promise<InvoiceItem | null> {
+  async findOne(
+    id: number,
+    invoiceId: number,
+    userId: number,
+  ): Promise<InvoiceItem | null> {
+    const invoice = await this.prisma.invoice.findUnique({
+      where: {
+        id: invoiceId,
+        userId,
+      },
+    });
+    if (!invoice) throw new NotFoundException(NOT_FOUND);
+
     const item = await this.prisma.invoiceItem.findUnique({
       where: {
         id,
@@ -40,7 +73,20 @@ export class ItemsService {
     return item;
   }
 
-  async update(id: number, updateItemDto: UpdateItemDto): Promise<InvoiceItem> {
+  async update(
+    id: number,
+    invoiceId: number,
+    userId: number,
+    updateItemDto: UpdateItemDto,
+  ): Promise<InvoiceItem> {
+    const invoice = await this.prisma.invoice.findUnique({
+      where: {
+        id: invoiceId,
+        userId,
+      },
+    });
+    if (!invoice) throw new NotFoundException(NOT_FOUND);
+
     const updatedItem = await this.prisma.invoiceItem.update({
       where: {
         id,
@@ -51,7 +97,19 @@ export class ItemsService {
     return updatedItem;
   }
 
-  async remove(id: number): Promise<InvoiceItem> {
+  async remove(
+    id: number,
+    invoiceId: number,
+    userId: number,
+  ): Promise<InvoiceItem> {
+    const invoice = await this.prisma.invoice.findUnique({
+      where: {
+        id: invoiceId,
+        userId,
+      },
+    });
+    if (!invoice) throw new NotFoundException(NOT_FOUND);
+
     const deletedItem = await this.prisma.invoiceItem.delete({
       where: {
         id,
